+++ src/pii_radar/integrations/__init__.py
@@ -1,0 +1,5 @@
+"""Azure Cloud Integration Package for pii-radar."""
+
+from pii_radar.integrations.azure import AzureBlobStreamRedactor, AzureEventHubHandler
+
+__all__ = ["AzureBlobStreamRedactor", "AzureEventHubHandler"]